[#22] Sql 결과 매핑 타입을 상위 모듈로 이동
